# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion de Comptes Bancaires",
        "description": "API pour la gestion des comptes bancaires avec authentification admin",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/admin/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister les comptes bancaires",
                "description": "Récupère la liste paginée des comptes avec possibilité de filtrage, tri et recherche",
                "operationId": "getComptes",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de compte (cheque/epargne)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "cheque",
                                "epargne"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Statut du compte (actif/inactif/bloque)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "inactif",
                                "bloque"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par numéro de compte ou nom du titulaire",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "dateCreation",
                            "enum": [
                                "dateCreation",
                                "solde",
                                "titulaire"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "uuid"
                                                    },
                                                    "numeroCompte": {
                                                        "type": "string",
                                                        "example": "COMP-20251025-EZ6TJPOP"
                                                    },
                                                    "titulaire": {
                                                        "type": "string",
                                                        "example": "Prof. Nicola Hessel"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "cheque",
                                                            "epargne"
                                                        ],
                                                        "example": "epargne"
                                                    },
                                                    "solde": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "452420.00"
                                                    },
                                                    "devise": {
                                                        "type": "string",
                                                        "example": "FCFA"
                                                    },
                                                    "dateCreation": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-10-25T19:21:35+00:00"
                                                    },
                                                    "statut": {
                                                        "type": "string",
                                                        "enum": [
                                                            "actif",
                                                            "inactif",
                                                            "bloque"
                                                        ],
                                                        "example": "actif"
                                                    },
                                                    "motifBlocage": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "metadata": {
                                                        "properties": {
                                                            "derniereModification": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "version": {
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "currentPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "totalPages": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "totalItems": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "itemsPerPage": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "hasNext": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "hasPrevious": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "properties": {
                                                "self": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8001/api/admin/comptes"
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8001/api/admin/comptes?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8001/api/admin/comptes?page=1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/client/login": {
            "post": {
                "tags": [
                    "Authentification Client"
                ],
                "summary": "Connexion client",
                "description": "Authentifie un client et retourne un token d'accès",
                "operationId": "clientLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "client@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "client": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "uuid"
                                                },
                                                "nom_complet": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "client@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Échec d'authentification",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/client/comptes": {
            "get": {
                "tags": [
                    "Comptes Client"
                ],
                "summary": "Lister les comptes du client",
                "description": "Récupère la liste des comptes du client connecté (comptes actifs de type cheque ou épargne)",
                "operationId": "getClientComptes",
                "responses": {
                    "200": {
                        "description": "Liste des comptes récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "uuid"
                                                    },
                                                    "numeroCompte": {
                                                        "type": "string",
                                                        "example": "COMP-20251025-EZ6TJPOP"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "cheque",
                                                            "epargne"
                                                        ],
                                                        "example": "epargne"
                                                    },
                                                    "solde": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "452420.00"
                                                    },
                                                    "devise": {
                                                        "type": "string",
                                                        "example": "FCFA"
                                                    },
                                                    "dateCreation": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-10-25T19:21:35+00:00"
                                                    },
                                                    "statut": {
                                                        "type": "string",
                                                        "enum": [
                                                            "actif",
                                                            "inactif",
                                                            "bloque"
                                                        ],
                                                        "example": "actif"
                                                    },
                                                    "motifBlocage": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}